Finished OfferTradeTest.java and testing on OfferTrade.java
Note, I didn’t test the cases where the user does not have the
resources necessary.  I want to know if that check happens in this
class or somewhere else.  @film42 did not implement this in the
AcceptTrade.java anyway so I’m led to believe that it does happen
somewhere else